feat(rust/signed-doc): Add CBOR deterministic `Map` as a separate struct, enable deterministic decoding for the metadata fields. (#396)
* move decode_map_deterministically impl into separate mod, provide a dedicated type Map

* cleanup

* Fix CBOR map minimal length validation timing

Capture decoder position before d.map() call to correctly read the map
header byte instead of the first key when validating minimal encoding.

Fixes test_map_minimal_length_encoding failure.

* add `DecodeCtx` struct

* wip

* wip

* wip

* replace BtreeMap with HashMap

* fixes

---------
