+++ rust/cbork-utils/src/lib.rs
@@ -1,5 +1,7 @@
 //! CBOR utility modules.
 
+pub mod decode_context;
 pub mod decode_helper;
 pub mod deterministic_helper;
+pub mod map;
 pub mod with_cbor_bytes;
